Added support for indirect invoke (functions as values).
A pre-pass creates a WFuncName constant (e.g. "func__g") for each
function 'g' whose address is taken, and a Boogie axiom is generated for
the possible application of each of those functions.  For example:
  axiom (forall arg1:WVal :: applyTo1(func_g, arg1) == g(arg1));
  
Then those constants are passed around as values, and are used as the
first argument of an 'applyTo<n>(func, args...)' function that calls the
function, with the above semantics.

This results in 12 more tests passing fully and 2 partially passing.

Current results:
$ wystats OUT36_indirect_invoke.txt 
Error Compiling:          21
NotImplementedYet:       109
Other Exceptions:          2
Errors in BPL:            12
Fully Verified:          346
Verify time outs:          3
Verify errors:            25